requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
langdetect==1.0.9
requests==2.31.0
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
psycopg2-binary==2.9.7
